Move SqlFunctionExecutor to SPI
SqlFunctionExecutor can be found by the classloader only when it is
accessible in SPI.
