+++ presto-function-namespace-managers/src/main/java/com/facebook/presto/functionNamespace/execution/thrift/SimpleAddressThriftSqlFunctionExecutionModule.java
@@ -16,9 +16,9 @@ package com.facebook.presto.functionNamespace.execution.thrift;
 import com.facebook.drift.client.address.SimpleAddressSelectorConfig;
 import com.facebook.presto.functionNamespace.execution.NoopSqlFunctionExecutor;
 import com.facebook.presto.functionNamespace.execution.SqlFunctionExecutionModule;
-import com.facebook.presto.functionNamespace.execution.SqlFunctionExecutor;
 import com.facebook.presto.spi.function.FunctionImplementationType;
 import com.facebook.presto.spi.function.RoutineCharacteristics.Language;
+import com.facebook.presto.spi.function.SqlFunctionExecutor;
 import com.facebook.presto.thrift.api.udf.ThriftUdfService;
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Binder;
