+++ presto-grpc-api/src/main/java/com/facebook/presto/grpc/GrpcSqlFunctionExecutor.java
@@ -19,7 +19,6 @@ import com.facebook.presto.common.block.BlockEncodingSerde;
 import com.facebook.presto.common.function.SqlFunctionResult;
 import com.facebook.presto.common.type.Type;
 import com.facebook.presto.common.type.TypeSignature;
-import com.facebook.presto.functionNamespace.execution.SqlFunctionExecutor;
 import com.facebook.presto.grpc.udf.GrpcFunctionHandle;
 import com.facebook.presto.grpc.udf.GrpcSerializedPage;
 import com.facebook.presto.grpc.udf.GrpcUdfInvokeGrpc;
@@ -30,6 +29,7 @@ import com.facebook.presto.grpc.udf.GrpcUdfResult;
 import com.facebook.presto.spi.function.FunctionImplementationType;
 import com.facebook.presto.spi.function.RemoteScalarFunctionImplementation;
 import com.facebook.presto.spi.function.RoutineCharacteristics.Language;
+import com.facebook.presto.spi.function.SqlFunctionExecutor;
 import com.facebook.presto.spi.function.SqlFunctionHandle;
 import com.facebook.presto.spi.function.SqlFunctionId;
 import com.google.inject.Inject;
